requirements.txt
requests>=2.28
python-dotenv>=1.0
google-generativeai
google-api-core
google-auth
pytest
